fix(investor,finance,sales): close TASK-19 security+review findings SR-1/SR-2, CR-1..CR-4, CR-6
- SR-1: RBAC owner/admin on data room mutations (POST /, /upload, DELETE)
  in investor/documents.ts - a viewer could permanently delete documents
- SR-2: reports/generate gated by owner/admin + Zod schema; byokKey/
  byokProvider are now enum/length-capped before reaching the Inngest job
- CR-1/CR-2: grant_date default and MRR month picker now use local
  calendar date instead of UTC (off-by-one near midnight for UTC-3 users;
  MRR could overwrite a neighbouring snapshot via UNIQUE)
- CR-3: equity DELETE /grants/:id maps Prisma P2025 to 404 (was 500)
- CR-4: .int() on options_total/total_options (fractional options rejected)
- CR-6: mrr_snapshots non-negative CHECK guard keyed by constraint name
  instead of any-CHECK-exists
- hygiene: dead a11y-audit bypass comment removed from DataRoomPage and
  InvestorUpdatesPage EOF
- tests: +7 new (RBAC 403s, byok validation, P2025 404, fractional 400)

CR-5 (roadmap_items status CHECK) still blocked: prod read denied, table
absent in beta. task.md TASK-19 statuses updated.

+++ apps/api/src/schemas/index.ts
@@ -229,7 +229,7 @@ export const updateInvestorUpdateSchema = z.object({
 export const createEquityGrantSchema = z.object({
   beneficiary_name: safeString(200),
   beneficiary_email: z.string().email().max(320).optional().nullable(),
-  options_total: z.number().positive().max(1_000_000_000),
+  options_total: z.number().int().positive().max(1_000_000_000),
   grant_date: z.string().max(100),
   cliff_months: z.number().int().min(0).max(120).default(12),
   vesting_months: z.number().int().positive().max(240).default(48),
@@ -242,7 +242,7 @@ export const createEquityGrantSchema = z.object({
 });
 
 export const upsertEsopPoolSchema = z.object({
-  total_options: z.number().positive().max(1_000_000_000),
+  total_options: z.number().int().positive().max(1_000_000_000),
   pool_percentage: z.number().positive().max(100).optional().nullable(),
   notes: safeText(2000).optional().nullable(),
 });
